exp(no-single-edm): fp32 diffusion forward with bf16 attention-core
New precision variant of the no-single qknorm EDM run: the bf16-autocast run
(epoch ~505) silently diverged (sampling lDDT 0.62 -> 0.25) while training loss
stayed flat. This experiment runs the diffusion forward in fp32 and keeps ONLY
the attention core in bf16, to test whether full-bf16 forward is the culprit.

- model.py: in the fp32 path (autocast_bf16=False) cast trunk conditioning
  (bf16) + x_t/t_emb to fp32 before the diffusion module.
- libs/team-gm: bump to attn-core-bf16-kernel (Triton kernel casts q/k/v/bias
  to bf16, restores caller dtype; no-op under bf16 autocast).
- config_exp_..._fp32attn_edm.yaml: identical to the bf16 config except
  model.diffusion.autocast_bf16=False (same data/LR/trunk-freeze/start ckpt).
- inference script: add --fp32, --attn-bf16, --save-traj flags used to
  reproduce and characterise the divergence.

Sanity (1-GPU training_step, fp32+attn-bf16): diffusion_loss 0.036,
total 0.107, grad_norm 212 (finite) -- matches bf16 baseline magnitude.

+++ src/miniworld/models/miniworld_no_single_at_trunk/model.py
@@ -426,15 +426,18 @@ class Model(nn.Module):
                 )
             # EuclideanDiffuser.cal_loss requires float32 x_update.
             return out.float()
+        # fp32 diffusion forward: the frozen trunk emits bf16 conditioning, so
+        # cast it (and x_t/t_emb) to fp32 here. The attention core still runs in
+        # bf16 inside the Triton kernel (see AugmentedAttentionPairBias).
         return self.diffusion_module(
             reference,
             scheme,
             structure,
-            x_t,
+            x_t.float(),
             x_mask,
-            t_emb,
-            token_single_input,
-            token_pair_trunk,
+            t_emb.float(),
+            token_single_input.float(),
+            token_pair_trunk.float(),
         )
 
     def forward(
